requirements.txt
torch>=1.11.0
ultralytics>=8.0.0
opencv-python
filterpy
lap
sort-tracker-py
scikit-image
numpy